Create Remove a node from BST
Problem statement
You are given a BST (Binary search tree) with’ N’ number of nodes and a node ‘X’ (which is definitely present in the input the BST). You need to remove the node ‘X’ and return the root of the updated BST.

For example:

 

Figure1 : Input BST and Figure2: Output BST

In the above example, For the given BST (Fig.1) and X = 6,  we have output( Fig. 2 ) after removing node 6.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 5
1 <= N <= 5 * 10 ^ 3
1 <= nodeVal[i] <= 10 ^ 9

Time Limit: 1 sec.
Sample Input 1:
1
10 5 15 2 6 -1 -1 -1 -1 -1 -1
6
Sample Output 1:
2 5 10 15
Explanation of Sample Input 1:
In the first test case, the input figure is given above in the input format explanation. 
After removing 6 from BST, the tree becomes as below:

So the inorder traversal of the updated tree is: 2 5 10 15
Sample Input 2:
1
2 1 3 -1 -1 -1 -1
2
Sample Output 2:
1 3
Explanation of Sample Output 2 :
 

Figure1: Input BST  and Figure2: Output BST

So the inorder traversal of the updated tree is: 1 3