Create Garden Fencing
Problem statement
Ninja owns a garden in the city of Ninjago. To prevent trespassers, he must fence the garden, but he is busy and asked you to help him.

You are given an array ‘Trees’ of size ‘N’ denoting the coordinates of each tree.

You want to impress Ninja by creating a circular fence such that all the trees lie either inside the fence or on the boundary of the fence. To minimize the cost you have to build the smallest possible fence. Find the coordinates of the centre and also find the radius of this circular fence.

Note :
Answers with a relative error of less than or equal to 10 ^ -5 are acceptable.
For Example :
If N = 3 and the trees are present at (0, 0), (1, 0) and (2, 0)
Then building a circle with (0, 0) and (2, 0) as diametric endpoints will result in the formation of the smallest circle enclosing all the trees.
Therefore we will return the centre as (1, 0) and radius equal to 1. 
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 ≤ T ≤ 10      
1 ≤ N ≤ 5000
-10^9 ≤ X, Y ≤ 10^9

Time limit: 1 sec
Sample Input 1 :
2
3
0 0
1 0
2 0
4
1 0
0 -1
-1 0
0 1
Sample Output 1 :
1 0 1
0 0 1
Explanation For Sample Input 1 :
For test case 1 :
Building a circular fence with (0, 0) and (2, 0) as diametric endpoints will result in the formation of the smallest circle enclosing all the trees.
Therefore we will return the centre as (1, 0) and radius equal to 1. 


For test case 2 : 
If we build a circular fence with a centre at (0, 0) and radius equal to 1, then all the trees will lie on the boundary of the fence.
Sample Input 2 :
2
6
1 1
2 0
2 2
2 4
3 3
4 2
6
1 1
0 2
2 2
2 4
3 3
4 2
Sample Output 2 :
2 2 2
2 2 2