Create Split Array Into The Subsets
Problem statement
Given an array, 'ARR' of 'N' integers print “true” if it is possible to split the array into some finite number of subsets such that each subset has the same integer in it and all subsets have the same size. The size of the subset should be greater than one.

Print “false” if it is not possible to do the above task.

Example:
Input: 'ARR' = [1, 2, 4, 4, 1, 2]
Output: true

We can split the above array like this: [1, 1], [4, 4], [2, 2] 
In the three subsets, each of size two. And every subset contains the same elements.

Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= ‘T’ <= 10
1 <= 'N' <= 10^5
-10^9 <= 'ARR[i]'<= 10^9

Time Limit : 1 sec
Sample Input 1 :
2
6
2 2 4 5 4 5
5
4 3 5 6 3
Sample Output 1 :
true
false
Explanation Of Sample Input 1 :
For the first test case, we can split the array like this: [2, 2], [4, 4], [5, 5].

For the second test case, it is not possible to split the array with the given condition.
Sample Input 2 :
2
6
2 3 4 5 3 5
9
2 3 3 2 3 1 2 1 1
Sample Output 2 :
false
true