Create Squares Sum
Problem statement
Given an array of ‘N’ integers and ‘q’ queries. The query is defined as below:-

Given two integers ‘l’ and ‘r’ ( ‘l’ >=0 and ‘r’ <= ‘N’) Find the sum of squares of all elements of the array with index in the range ‘l’ and ‘r’ (Inclusive).

Given three integers ‘l’ ,’r’ and ‘x’ ( ‘l’ >=0 and ‘r’ <= ‘N’) .Update the elements of the array by incrementing ‘x’ to all array elements with index in the range ‘l’ and ‘r’.

Given three integers ‘l’ ,’r’ and ‘x’ ( ‘l’ >=0 and ‘r’ <= ‘N’) .Update the elements of the array by setting ‘x’ to all array elements with index in the range ‘l’ and ‘r’.

Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= ’T’ <= 10
1 <= ‘N’ <= 1000
1 <= q <= 1000
1 <= arr[i] <= 1000
1 <= ‘x’ <= 1000

Where ‘i’ varies from 0 to ‘n’ -1  where ‘n’ is the length of the array ‘arr’,q is the number of queries ‘arr[i]’ is the value of the element at index ‘i’. ‘x’ is the number for updating the range. 


Time Limit: 1 sec
Sample Input 1 :
1
5
1 2 3 4 5
5
2 0 2
0 0 2 1
2 0 3
1 0 2 2
2 0 4
Sample Output 1 :
14
45
53
Sample Output Explanation :
The first line has a single integer which means it’s a single test case.
The next line has an integer 5 denoting that there are 5 integers in the array.
Then the 5 space-separated integers in the third line represent the elements in the array.
Then the fourth line has ‘q’ which is the number of queries. In this case, we have 5 queries.

The 5 queries are as follow:-

In first query type = 2   ‘l’ = 0 and ‘r’ = 2
The sum of squares is  arr[0]^2 + arr[1]^2 + arr[2]^2 = 1^2 + 2^2 + 3^2 = 14

In second query type = 0  ‘l’ = 0 and ‘r’ = 2 and ‘x’ = 1.
Incrementing ‘x’ from 0 to 2 in an array. Thus array becomes [ 2,3,4,4,5].

In third query type = 2 ‘l’ = 0 and ‘r’ = 3.
The sum of squares is arr[0]^2 + arr[1]^2 + arr[2]^2 + arr[3]^3 = 2^2 + 3^2 + 4^2 + 4^2 = 45

In fourth query type = 1  ‘l’ = 0 and ‘r’ = 2 and ‘x’ = 2.
Setting ‘x’ from 0 to 2 in an array. Thus array becomes [2,2,2,4,5].

In fifth query type = 2   ‘l’ = 0 and ‘r’ = 4
The sum of squares is arr[0]^2 + arr[1]^2 + arr[2]^2 + arr[3]^3 + arr[3]^2  = 53
Sample Input 2 :
2
5
1 10 5 1 3 
4
0 1 1 10
1 2 2 6
0 1 4 6
2 0 4 
6
6 2 8 3 5 2 
3
0 1 2 2
2 1 1 
2 0 2  
Sample Output 2 :
951
16
152